+++ src/main/java/org/fugerit/java/client/redis/ClientRedisArgs.java
@@ -1,0 +1,16 @@
+package org.fugerit.java.client.redis;
+
+public class ClientRedisArgs {
+
+	public static final String ARG_MODE = "mode";
+	public static final String MODE_SINGLE_COMMAND = "single-command";
+	public static final String MODE_GUI = "gui";
+	public static final String MODE_DEFAULT = MODE_SINGLE_COMMAND;
+	
+	public static final String ARG_REDIS_URL = "redis-url";
+	public static final String ARG_KEY = "key";
+	public static final String ARG_VALUE = "value";
+	public static final String ARG_TTL = "ttl";
+	
+}
+