+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 # Changelog for project fj-client-redis
 
 ## version 0.1.0 - XXXX-XX-XX
-* Simple connection test for redis
+* Support for set / get of a single key at command line